Alter auto-initdb to only happen for server/submit on sqlite
Before we had the database always connect if configured and
ensure the schema existed. This is a problem if the database
is not available on the client side (which should be okay).

Instead, we only auto-initdb on sqlite databases on
server and submit workflow startup. And for full databases
we have added an `initdb` command.