[lib] Remove config file permissions check
This feature was a bit short-sighted. The goal was to add a layer of
protection for users that might add sensitive credentials to the config
file. However, this is not necessarily meaningful on all platforms, and
so hard to implement, and we provide other mechanisms to protect data
(i.e., _env and _eval fields).

Because we haven't made a release yet since this was introduced, we will
simply increment the minor version.