[lib] Fix shared exception mapping (#16)
The top-level override of the base class `Application.exceptions` does not
actually work because we've already imported all the end `Application`
implementations (so the `{..., **Application.exceptions}` stuff does nothing).
We've now moved the dictionary into a method call with no dependencies within
the `core` subpackage, so it can be imported appropriately everywhere.