[lib] Fix client update when not using database
The new `Client.add` and `Client.update` calls by the server-side
heartbeat monitor needed to be contained within checks on if we were
in-memory or using the database.

The updates were harmless but output in logs would be confusing and
pointless because the in-memory db goes away on program shutdown.