[lib] Add AutoScaling support (and client timeout)
Autoscaling cluster allows for elastically scaling up and down
based on task throughput and pressure. Enabled with --autoscaling,
and implemented with corresponding environment variables and CLI
options.

This feature is well tested but still someone experimental.
Should we add an experimental flag or variable to enable/disable
new features like this?

Additional missing bits and fixes for client --timeout and such
are included in this commit because they were entangled.