Add new 'ready' feature to server
Previously we just slept for 2 seconds to allow the server to
finish starting up and avoid a client crash because we were
too quick.

This was lazy and not technically correct because the server
could have taken longer in the pathological case. This new
implementation sets a flag that we wait for to proceed.
Total runtime of the application in the no-task limiting
case is now less than a second!

Signed-off-by: Geoffrey Lentner <glentner@purdue.edu>